Extract findParamMatchingId() and add docs
